user_service/requirements.txt
fastapi
uvicorn

sqlalchemy
asyncpg
alembic

python-jose[cryptography]
passlib[bcrypt]

email-validator
python-multipart
bcrypt